+++ crypto/src/main/java/pm/gnosis/crypto/utils/CryptoExtensions.kt
@@ -2,6 +2,9 @@ package pm.gnosis.crypto.utils
 
 import okio.ByteString
 import org.spongycastle.jcajce.provider.digest.RIPEMD160
+import pm.gnosis.model.Solidity
+import pm.gnosis.utils.asEthereumAddressString
+import pm.gnosis.utils.removeHexPrefix
 import java.lang.AssertionError
 import java.security.NoSuchAlgorithmException
 
@@ -13,5 +16,20 @@ fun ByteString.hash160(): ByteString {
     } catch (e: NoSuchAlgorithmException) {
         throw AssertionError(e)
     }
-
 }
+
+fun Solidity.Address.asEthereumAddressChecksumString() =
+    asEthereumAddressString().removeHexPrefix().run {
+        val checksum = Sha3Utils.keccak(toByteArray())
+        foldIndexed(StringBuilder("0x"), { index, stringBuilder, char ->
+            stringBuilder.append(
+                when {
+                    char in '0'..'9' -> char
+                    checksum.hexCharValue(index) >= 8 -> char.toUpperCase()
+                    else -> char.toLowerCase()
+                }
+            )
+        }).toString()
+    }
+
+private fun ByteArray.hexCharValue(position: Int) = (get(position / 2).toInt() ushr (4 * ((position + 1) % 2))) and 0x0F
